pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend base para Viñas AI"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
supabase = "^2.5.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"